ci: add meson permutation in the Ubuntu builds
A number of Linux distributions have switched to the meson build. In
particular - Arch, Fedora, Gentoo and more are to follow in due time.

Even the libva builds that Chrome uses (aka instrumented libraries) is
using meson (on top of Ubuntu, currently aiming at switching to Focal).

Note: meson 0.53 (as used in Ubuntu 20.04) lack the `meson compile`, so
we fallback to calling `ninja` directly in that case.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>